@@ -29,7 +29,7 @@ classifiers = [
 ]
 dynamic = ["version"]
 
-requires-python = ">=3.7"
+requires-python = ">=3.8"
 dependencies = [
     "numpy~=1.20",
     "scipy>=1.5.3, <2",
@@ -37,12 +37,12 @@ dependencies = [
     "plum-dispatch>=1.5.1, <3",
     "numba>=0.52, <0.58",
     "igraph>=0.9.8, <0.11.0",
-    "jax>=0.3.16, <0.5",
-    "jaxlib>=0.3.15, <0.5",
-    "flax>=0.6, <0.7",
+    "jax>=0.4.2, <0.5",
+    "jaxlib>=0.4.2, <0.5",
+    "flax>=0.6.5, <0.7",
     "orjson>=3.4, <4",
     "optax>=0.1.3, <0.2",
-    "numba4jax>=0.0.10, <0.1",
+    "numba4jax>=0.0.10.post1, <0.1",
 ]
 
 [project.optional-dependencies]
