@@ -151,7 +151,7 @@ class SteadyState(AbstractVariationalDriver):
 
     def info(self, depth=0):
         lines = [
-            "{}: {}".format(name, info(obj, depth=depth + 1))
+            f"{name}: {info(obj, depth=depth + 1)}"
             for name, obj in [
                 ("Lindbladian ", self._lind),
                 ("Optimizer   ", self._optimizer),
