@@ -127,7 +127,7 @@ class TDVP(TDVPBaseDriver):
 
     def info(self, depth=0):
         lines = [
-            "{}: {}".format(name, info(obj, depth=depth + 1))
+            f"{name}: {info(obj, depth=depth + 1)}"
             for name, obj in [
                 ("generator     ", self._generator_repr),
                 ("integrator    ", self._integrator),
