@@ -34,7 +34,7 @@ def bool_env(varname: str, default: bool) -> bool:
     elif val in ("n", "no", "f", "false", "off", "0"):
         return False
     else:
-        raise ValueError("invalid truth value %r for environment %r" % (val, varname))
+        raise ValueError(f"invalid truth value {val!r} for environment {varname!r}")
 
 
 def int_env(varname: str, default: int) -> int:
