@@ -18,7 +18,6 @@ from jax import numpy as jnp
 
 from netket.hilbert import Fock
 from netket.utils.dispatch import dispatch
-from netket.utils import pure_callback
 
 
 @dispatch
@@ -31,7 +30,7 @@ def random_state(hilb: Fock, key, batches: int, *, dtype=np.float32):
         return jnp.asarray(rs, dtype=dtype)
 
     else:
-        state = pure_callback(
+        state = jax.pure_callback(
             lambda rng: _random_states_with_constraint(hilb, rng, batches, dtype),
             jax.ShapeDtypeStruct(shape, dtype),
             key,
