@@ -13,22 +13,10 @@
 # limitations under the License.
 
 import numpy as np
-
-from netket.utils import module_version
+import jax
 
 from .types import Array, DType, Shape
 from .struct import dataclass
-
-# TODO keep only  jax>=0.4 is required
-if module_version("jax") >= (0, 4, 0):
-    import jax
-
-    JaxArray = jax.Array
-else:
-    # pre jax 0.4
-    import jaxlib
-
-    JaxArray = jaxlib.xla_extension.DeviceArray
 
 
 @dataclass(cache_hash=True)
@@ -48,7 +36,7 @@ class HashableArray:
         if isinstance(wrapped, HashableArray):
             wrapped = wrapped.wrapped
         else:
-            if isinstance(wrapped, JaxArray):
+            if isinstance(wrapped, jax.Array):
                 # __array__ only works if it's a numpy array.
                 wrapped = np.array(wrapped)
             else:
